fix(): Break the import cycle 2
in the telegram namespace we have a cycle types.ts <-> text/index.ts. We extract the commands into separate file so solve this
