+++ src/telegram/actions/lang.ts
@@ -1,10 +1,10 @@
 import { GenericAction } from "./common.js";
 import {
-  BotCommand,
   BotLangData,
   TelegramButtonModel,
   TelegramMessagePrefix,
 } from "../types.js";
+import { BotCommand } from "../commands.js";
 import { BotMessageModel } from "../model.js";
 import {
   getLanguageByText,
