+++ src/telegram/actions/start.ts
@@ -3,7 +3,8 @@ import { isCommandMessage } from "../helpers.js";
 import { TranslationKeys } from "../../text/types.js";
 import { Logger } from "../../logger/index.js";
 import { collectAnalytics } from "../../analytics/index.js";
-import { BotCommand, type TelegramMessagePrefix } from "../types.js";
+import { type TelegramMessagePrefix } from "../types.js";
+import { BotCommand } from "../commands.js";
 import type { BotMessageModel } from "../model.js";
 import type { TgMessage } from "../api/types.js";
