+++ src/telegram/actions/support.ts
@@ -4,7 +4,8 @@ import { TranslationKeys } from "../../text/types.js";
 import { githubUrl, officialChannelAccount } from "../../const.js";
 import { Logger } from "../../logger/index.js";
 import { collectAnalytics } from "../../analytics/index.js";
-import { BotCommand, type TelegramMessagePrefix } from "../types.js";
+import { type TelegramMessagePrefix } from "../types.js";
+import { BotCommand } from "../commands.js";
 import type { BotMessageModel } from "../model.js";
 import type { TgMessage } from "../api/types.js";
 import type { TgInlineKeyboardButton } from "../api/groups/chats/chats-types.js";
